Fix "podcasts" redirection
If a page with the slug "podcasts" existed WordPress did not know which rewrite rule to use. It checks now if a page with that slug is active. If yes the link will point to the page and not the plugin archive. If no site exists with that slug the archive is generated.